Some Exceptions and tests
