+++ Particle_Simulation/test/test_Parameters.py
@@ -1,5 +1,6 @@
 from Particle_Simulation.Parameters import Parameters
 import numpy as np
+import numpy.testing as npt
 import unittest
 
 
@@ -7,22 +8,22 @@ class test_Parameters(unittest.TestCase):
     def test_3Dkvector(self):
         Parameter = Parameters(temperature=1, box=[1, 1, 1], es_sigma=2, update_radius=0.5, cutoff_radius=1, K_cutoff=2,
                                particle_types=np.array([0]))
-        reference = [[-2, 0, 0], [-1, -1, -1], [-1, -1, 0], [-1, -1, 1], [-1, 0, -1], [-1, 0, 0], [-1, 0, 1],
-                     [-1, 1, -1], [-1, 1, 0], [-1, 1, 1], [0, -2, 0], [0, -1, -1], [0, -1, 0], [0, -1, 1], [0, 0, -2],
-                     [0, 0, -1]]
-        self.assertEqual(Parameter.k_vector, reference)
+        reference = np.array([[-2, 0, 0], [-1, -1, -1], [-1, -1, 0], [-1, -1, 1], [-1, 0, -1], [-1, 0, 0], [-1, 0, 1],
+                              [-1, 1, -1], [-1, 1, 0], [-1, 1, 1], [0, -2, 0], [0, -1, -1], [0, -1, 0], [0, -1, 1], [0, 0, -2],
+                              [0, 0, -1]])
+        npt.assert_array_equal(Parameter.k_vector, reference)
 
     def test_2Dkvector(self):
-        Parameter = Parameters(temperature=1, box=[1, 1, 1], es_sigma=2, update_radius=0.5, cutoff_radius=1, K_cutoff=2,
+        Parameter = Parameters(temperature=1, box=[1, 1], es_sigma=2, update_radius=0.5, cutoff_radius=1, K_cutoff=2,
                                particle_types=np.array([0]))
-        reference = [[-2, 0], [-1, -1], [-1, 0], [-1, 1], [0, -2], [0, -1]]
-        self.assertEqual(Parameter.k_vector, reference)
+        reference = np.array([[-2, 0], [-1, -1], [-1, 0], [-1, 1], [0, -2], [0, -1]])
+        npt.assert_array_equal(Parameter.k_vector, reference)
 
     def test_1Dkvector(self):
         Parameter = Parameters(temperature=1, box=[1], es_sigma=2, update_radius=0.5, cutoff_radius=1, K_cutoff=2,
                                particle_types=np.array([0]))
-        reference = [[-2], [-1]]
-        self.assertEqual(Parameter.k_vector, reference)
+        reference = np.array([[-2], [-1]])
+        npt.assert_array_equal(Parameter.k_vector, reference)
 
     def test_negative_temperature(self):
         self.assertRaises(ValueError, Parameters, -2, [1, 2, 3], 2, 2, np.array([0]), 2, 2)
